% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocate.R
\name{colocate}
\alias{colocate}
\title{Function to colocate gauges with corresponding GPM pixel}
\usage{
colocate(gauges, gpm, resolution, longlat = TRUE)
}
\arguments{
\item{gauges}{A STFDF object.}

\item{gpm}{a STFDF object.}

\item{resolution}{An integer describing the satellite spatial resolution.}

\item{longlat}{A logical object indicating if object is in long/lat or not. TRUE by default.}
}
\value{
Object \code{gauges} with additional spatial column indicating GPM pixel ID.
}
\description{
xxx
}
\details{
specifics of calculation (uses spdistsn1)
}
\examples{
#add(1, 1)
#add(10, 1)
}

