% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_download.R
\name{system_download}
\alias{system_download}
\title{Function to download GPM-IMERG HDF5 files from NASA PPS}
\usage{
system_download(product, nrt_product, OS)
}
\arguments{
\item{product}{A character string describing which GPM product is selected}

\item{nrt_product}{A character string describing whih NRT GPM product is used.}

\item{OS}{A character string describing the operating system.}
}
\value{
Downloads and writes HDF files. No R object created.
}
\description{
xxx
}
\details{
requires file list of URLs to download HDF files using wget/ curl.
}
\examples{
#add(1, 1)
#add(10, 1)
}

